requirements.txt
rank-bm25==0.2.2
mcp
opentelemetry-sdk
requests
